requirements.txt
feedparser>=6.0.0
requests>=2.28.0
schedule>=1.2.0
python-dotenv>=1.0.0
transformers>=4.30.0
email-validator>=2.0.0
Pillow>=9.0.0
torch>=1.10.0
numpy==1.24.3
streamlit>=1.24.0
pandas>=1.5.0
